+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Added iOS support
 - Added early loading
 - Added curl and info copy shortcuts
+- Moved theme buttons to settings
 - Moved keybind settings to the GDIntercept settings
 - Made Geode proxies less intrusive in the client
 - Fixed incompatibilities with other Geode web hooking mods